% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age_dist.R
\name{age_dist}
\alias{age_dist}
\title{Create distance from age column}
\usage{
age_dist(dat, extra_column = c(mo = "months", yr = "years"))
}
\arguments{
\item{dat}{a vector containing ages in years, or, if a data frame, the first
column contains the ages and the second contains the specifier of type of
age (months or years).}

\item{extra_column}{a named vector specifying the type of age represented.}
}
\value{
a distance matrix of differences in age.
}
\description{
Create distance from age column
}
\examples{
x <- rpois(5, 30)
y <- rpois(5, 18)
dat <- data.frame(age = c(x, y), age_class = rep(c("YEAR", "MONTH"), each = 5))
age_dist(x)
age_dist(dat, extra_column = c(mo = "MONTH", yr = "YEAR"))
}

