% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchAge.R
\name{ageDists}
\alias{ageDists}
\title{Create distance from age column}
\usage{
ageDists(dat1, dat2 = NULL, e = 1, extra_column = c(mo = "months", yr =
  "years"))
}
\arguments{
\item{dat1}{a data frame of ages}

\item{dat2}{a data frame of ages}

\item{e}{the number of years difference allowed for two entries to be the same}

\item{extra_column}{a named vector specifying the type of age represented}
}
\value{
a distance matrix of differences in age.
}
\description{
This function will take one or two data frames containing ages
  and create a distance matrix based off of the manhattan distance. Data can
  be entered as a single column data frame assuming years or a two column
  data frame with a specifier for months or year in the second column.
}
\examples{
set.seed(999)
x <- data.frame(age = rpois(5, 30), age_class = "YEAR")
y <- data.frame(age = rpois(5, 18), age_class = "MONTH")
dat <- rbind(x, y)
ageDists(dat[1])
ageDists(dat, extra_column = c(mo = "MONTH", yr = "YEAR"))
ageDists(x[1], y, extra_column = c(mo = "MONTH", yr = "YEAR"))
}

