% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapseDistMatrices.R
\name{collapseDistMatrices}
\alias{collapseDistMatrices}
\title{Collapse/sum matrices using weights}
\usage{
collapseDistMatrices(d1, d2, weightVector = c(0.5, 0.5))
}
\arguments{
\item{list}{of distances matrices and a vector of corresponding weights
default: all equal weights for each input matrix}
}
\value{
a single weighted distance matrix
}
\description{
Collapse/sum matrices using weights
}
\examples{
just sum two of the same matrices:
caseIDVector <- c("AB-10-1", "AB-10-5", "AB-10_1")
d1 <- exactMatchCaseIDIntraDataset(caseIDVector)
d2 <- exactMatchCaseIDIntraDataset(caseIDVector)
summedDistMatrix <- collapseDistMatrices(d1, d2, c(0.5, 0.5))
}

