% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distFuns.R
\name{distFuns}
\alias{distFuns}
\title{Get all the distance functions in epimatch}
\usage{
distFuns()
}
\value{
a named vector of distance functions in epimatch
}
\description{
Get all the distance functions in epimatch
}
\examples{
distFuns()
}

