% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi_self_match.R
\name{epi_self_match}
\alias{epi_self_match}
\title{Find duplicates in single data set}
\usage{
epi_self_match(dat, funlist = list(name = "match_names", age =
  list("age_list", extra_column = c(mo = "months", yr = "years")), CaseID =
  "exactMatchCaseIDIntraDataset"))
}
\arguments{
\item{dat}{a linelist}

\item{funlist}{a list of functions and arguments to use to evaluate columns of dat.}
}
\description{
Find duplicates in single data set
}

