% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exactMatchCaseID.R
\name{exactMatchCaseIDIntraDataset}
\alias{exactMatchCaseIDIntraDataset}
\title{match case ID given a vector of caseIDs}
\usage{
exactMatchCaseIDIntraDataset(caseIDVector)
}
\arguments{
\item{caseIDVector}{a vector of caseIDs}
}
\value{
a distance matrix
}
\description{
match case ID given a vector of caseIDs
}
\examples{
caseIDVector <- c("AB-10-1", "AB-10-5", "AB-10_1")
exactMatchCaseIDIntraDataset(caseIDVector)
}

