% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchGender.R
\name{genderDists}
\alias{genderDists}
\title{gender function for matching
A thin wrapper around generic, which converts some known male and female
codes to binary}
\usage{
genderDists(dat1, dat2 = NULL, m = c("m", "male", "mal", "1"), f = c("f",
  "female", "fem", "2"))
}
\arguments{
\item{dat1}{a vector of genders to match}

\item{dat2}{a vector of genders to match}

\item{m}{format of males}

\item{f}{format of females}
}
\value{
a pairwise matrix of scores from 0 (exact match) to 1 (no match)
}
\description{
gender function for matching
A thin wrapper around generic, which converts some known male and female
codes to binary
}
\examples{
test <- data.frame(c("male", "f", "m"), stringsAsFactors = FALSE)
genderDists(test)
}

