% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{generic_match}
\alias{generic_match}
\title{generic function for matching}
\usage{
generic_match(dat, ...)
}
\arguments{
\item{dat}{a vector of data to determine matching}

\item{...condiditions}{to apply}
}
\value{
a pairwise matrix of scores from 0 (exact match) to 1 (no match)
}
\description{
generic function for matching
}
\examples{
generic_match(1:3)
}

