% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchGeneric.R
\name{genericDists}
\alias{genericDists}
\title{generic function for exact matching single columns}
\usage{
genericDists(dat1, dat2 = NULL, strings = F)
}
\arguments{
\item{dat1}{a vector of data to determine matching}

\item{dat2}{a vector of data to determine matching}

\item{strings}{whether to use the edit distance rather than exact matches}
}
\value{
a pairwise matrix of scores from 0 (exact match) to 1 (no match)
}
\description{
generic function for exact matching single columns
}
\examples{
generic_match(1:3)
}

