% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi_self_match.R
\name{getIndexList}
\alias{getIndexList}
\title{return a list of indices from matchEpiData}
\usage{
getIndexList(matchList)
}
\arguments{
\item{matchList}{a list of lists of weights}
}
\value{
an unnamed list of lists of indices
}
\description{
The output of matchEpiData can either be a list of indices or a named list of
weights. If it's a named list of weights, this function will return the list
of indices.
}
\examples{
## Loading Data
library('epimatch')
indata <- system.file("files", package = "epimatch")
indata <- dir(indata, full.names = TRUE)
x <- lapply(indata, read.csv, stringsAsFactors = FALSE)
names(x) <- basename(indata)

# We will use one data set from the case information and lab results
case <- x[["CaseInformationForm.csv"]]
lab <- x[["LaboratoryResultsForm7.csv"]]

funlist <- list(
             list(d1vars = "ID",
                  d2vars = "ID",
                  fun = "nameDists",
                  extraparams = NULL,
                  weight = 1),
             list(d1vars = c("Surname", "OtherNames"),
                  d2vars = c("SurnameLab", "OtherNameLab"),
                  fun = "nameDists",
                  extraparams = NULL,
                  weight = 0.5)
           )
# This will get all of the indices that match the ID and Names with a
# threshold of 0.25
res <- matchEpiData(dat1 = case,
                    dat2 = lab,
                    funlist = funlist,
                    thresh = 0.25,
                    giveWeight = TRUE)
res               # List of weights
getIndexList(res) # List of indices
}
\keyword{internal,}
\keyword{utilities}

