% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchLocation.R
\name{locationDists}
\alias{locationDists}
\title{Compare locations}
\usage{
locationDists(a, b)
}
\arguments{
\item{a}{data frame containing location information}

\item{b}{data frame containing location information}
}
\value{
a disttance matrix
}
\description{
Compare locations
}
\examples{
test <- data.frame(country=c("Uganda", "Uganda", "Kenya"),
                   city=c("Kampala", "Kampala", "Kilifi"), stringsAsFactors = F)
test2 <- data.frame(location=c("uganda,kampala", "uganda,kampala", "kenya"),
stringsAsFactors = F)
}

