% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_names.R
\name{match_names}
\alias{match_names}
\title{Clean and match names column by removing punctuation and replacing
punctuation with an underscore}
\usage{
match_names(dat)
}
\arguments{
\item{dat}{a vector or data frame}
}
\value{
a distance matrix giving the number of mismatches in names.
}
\description{
Clean and match names column by removing punctuation and replacing
punctuation with an underscore
}
\examples{
set.seed(9)
x <- data.frame(x = letters, y = LETTERS, z = 1:26)
x <- sample(nrow(x), 10, replace = TRUE)
match_names(x)
}

