% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchNames.R
\name{nameDists}
\alias{nameDists}
\title{Clean and match names column by removing punctuation and replacing
punctuation with an underscore}
\usage{
nameDists(dat1, dat2 = NULL)
}
\arguments{
\item{dat1}{data frame containing name information}

\item{dat2}{data frame containing name information}
}
\value{
a distance matrix giving the number of mismatches in names.
}
\description{
Clean and match names column by removing punctuation and replacing
punctuation with an underscore
}
\examples{

set.seed(9)
x <- data.frame(x = letters, y = LETTERS, z = 1:26)
x <- x[sample(nrow(x), 10, replace = TRUE), ]
nameDists(x)
}

