% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_matching.R
\name{processFunctionList}
\alias{processFunctionList}
\title{Process the functions used for matching data}
\usage{
processFunctionList(dat1, dat2 = NULL, funlist = list())
}
\arguments{
\item{dat1}{An input linelist}

\item{dat2}{An optional extra linelist}

\item{funlist}{A list containing lists containing:
\itemize{
 \item d1vars - variable names for dataset 1
 \item d2vars - variable names for dataset 2
 \item fun - function name to process on these variables
 \item extraparams - extra parameters that need to be applied with the function.
 \item weights - a weight vector to scale each matrix (not used in processFunctionList).
}}
}
\value{
a list of distance matrices scaled by weight
}
\description{
Input one or two data sets, process them with a list of containing functions,
column names, and function parameters, and return a list of matrices by
function.
}
\examples{
indata <- system.file("files", package = "epimatch")
indata <- dir(indata, full.names = TRUE)
x      <- lapply(indata, read.csv, stringsAsFactors = FALSE)

processFunctionList(dat1 = x[[1]],
                 dat2 = NULL,
                 funlist = list(
                   ID = list(d1vars = "Outbreak.ID.",
                             d2vars = NULL,
                             fun = "nameDists",
                             extraparams = NULL,
                             weight = 0.5),
                   names = list(d1vars = "Name..as.given.",
                                d2vars = NULL,
                                fun = "nameDists",
                                extraparams = NULL,
                                weight = 0.5),
                   list(d1vars = c("Age", "AgeUnit"),
                        d2vars = NULL,
                        fun = "ageDists",
                        extraparams = list(e = 5),
                        weight = 0.5)
                   )
                 )

}

