% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapseDistMatrices.R
\name{collapseDistMatrices}
\alias{collapseDistMatrices}
\title{Collapse/sum matrices using weights}
\usage{
collapseDistMatrices(distMatrixList, weightVector = NULL)
}
\arguments{
\item{distMatrixList}{list of distances matrices and a vector of
corresponding weights default: all equal weights for each input matrix}

\item{weightVector}{a vector of weights for each matrix. If \code{NULL}, then
each matrix will be weighted equally.}
}
\value{
a single weighted distance matrix
}
\description{
Collapse/sum matrices using weights
}
\examples{
# just sum two of the same matrices:
caseIDVector <- c("AB-10-1", "AB-10-5", "AB-10_1")
m1 <- exactMatchCaseIDIntraDataset(caseIDVector)
m2 <- exactMatchCaseIDIntraDataset(caseIDVector)
summedDistMatrix <- collapseDistMatrices(list(m1, m2), c(0.5, 0.5))
}

