% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapseDistMatrices.R
\name{collapseDistMatrices}
\alias{collapseDistMatrices}
\title{Collapse/sum matrices using weights}
\usage{
collapseDistMatrices(distMatrixList, weightVector = NULL, NAweight = 0.5)
}
\arguments{
\item{distMatrixList}{list of distances matrices and a vector of
corresponding weights default: all equal weights for each input matrix}

\item{weightVector}{a vector of weights for each matrix. If \code{NULL}, then
each matrix will be weighted equally.}

\item{NAweight}{a custom parameter specifying the score to be assigned for
missing data.}
}
\value{
a single weighted distance matrix
}
\description{
Collapse/sum matrices using weights
}
\examples{
# just sum two of the same matrices:
caseIDVector <- data.frame(c("AB-10-1", "AB-10-5", "AB-10_1"))
m1 <- nameDists(caseIDVector)
m2 <- nameDists(caseIDVector)
summedDistMatrix <- collapseDistMatrices(list(m1, m2), c(0.5, 0.5))
}

