% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchDate.R
\name{dateDists}
\alias{dateDists}
\title{date function for matching}
\usage{
dateDists(dat1, dat2 = NULL, dat1Format = "\%d-\%b-\%y",
  dat2Format = "\%d-\%b-\%y", threshold = NULL)
}
\arguments{
\item{dat1}{a vector of dates to match}

\item{dat2}{a vector of dates to match}

\item{dat1Format}{format of dat1 dates}

\item{dat2Format}{format of dat2 dates}

\item{threshold}{in days, otherwise returns scaled values}
}
\value{
a pairwise matrix of scores from 0 (exact match) to 1 (no match)
}
\description{
date function for matching
}
\examples{
test <- data.frame(c("21-Jan-01", "25-Jan-02", "21-Jan-01"), stringsAsFactors = F)
dateDists(test)
}

