% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchDate.R
\name{dateDists}
\alias{dateDists}
\title{date function for matching}
\usage{
dateDists(dat1, dat2 = NULL, dat1Format = "mdy", dat2Format = "mdy",
  threshold = NULL)
}
\arguments{
\item{dat1}{a vector of dates to match}

\item{dat2}{a vector of dates to match}

\item{dat1Format}{format of dat1 dates to pass to \pkg{lubridate}}

\item{dat2Format}{format of dat2 dates to pass to \pkg{lubridate}}

\item{threshold}{in days, otherwise returns scaled values}
}
\value{
a pairwise matrix of scores from 0 (exact match) to 1 (no match)
}
\description{
date function for matching
}
\details{
The formats for dates are passed to \pkg{lubridate}'s function
  \code{\link{parse_date_time}}, so they need to be in the format "mdy" or
  "ymd", etc.
}
\examples{
test <- c("Jan-21-01", "Jan-25-02", "Jan-21-01", "Jan-31-01")
test <- data.frame(test, stringsAsFactors = FALSE)

# match dates within 11 days of each other
dateDists(test, threshold = 11)
}

