% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi_self_match.R
\name{epi_self_match}
\alias{epi_self_match}
\title{Find duplicates in single data set}
\usage{
epi_self_match(dat1, dat2 = NULL, funlist = list())
}
\arguments{
\item{dat1}{An input linelist}

\item{dat2}{An optional extra linelist}

\item{funlist}{A list containing lists containing:
\itemize{
 \item d1vars - variable names for dataset 1
 \item d2vars - variable names for dataset 2
 \item fun - function name to process on these variables
 \item extraparams - extra parameters that need to be applied with the function.
 \item weights - a weight vector to scale each matrix (not used in process_matching).
}}
}
\value{
something
}
\description{
Find duplicates in single data set
}
\details{
this function will take in one or two data sets and also a
}
\examples{
rnorm(10)
}

