% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchGeneric.R
\name{genericDists}
\alias{genericDists}
\title{generic function for exact matching single columns}
\usage{
genericDists(dat1, dat2 = NULL, strings = FALSE)
}
\arguments{
\item{dat1, dat2}{data frames of data to match}

\item{strings}{whether to use the edit distance rather than exact matches}
}
\value{
a pairwise matrix of scores from 0 (exact match) to 1 (no match)
}
\description{
generic function for exact matching single columns
}
\examples{
library('epimatch')
set.seed(9)
x <- data.frame(dat = sample(10, replace = TRUE))
x$let <- letters[x$dat]
genericDists(x["dat"])
genericDists(x["let"], strings = TRUE)
}

