% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchLocation.R
\name{locationDists}
\alias{locationDists}
\title{Compare locations}
\usage{
locationDists(dat1, dat2 = NULL)
}
\arguments{
\item{dat1}{data frame containing location information}

\item{dat2}{data frame containing location information}
}
\value{
a distance matrix
}
\description{
Compare locations
}
\examples{
test <- data.frame(country=c("Uganda", "Uganda", "Kenya"),
                   city=c("Kampala", "Kampala", "Kilifi"),
                   stringsAsFactors = FALSE)
test2 <- data.frame(location=c("uganda,kampala", "uganda,kampala", "kenya"),
                    stringsAsFactors = FALSE)
locationDists(test, test2)
}

