% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi_self_match.R
\name{matchEpiData}
\alias{matchEpiData}
\title{Find duplicates in one or two data sets}
\usage{
matchEpiData(dat1, dat2 = NULL, funlist = list(), thresh = 0.05,
  giveWeight = FALSE)
}
\arguments{
\item{dat1}{An input linelist}

\item{dat2}{An optional extra linelist}

\item{funlist}{A list containing lists containing:
\itemize{
 \item d1vars - variable names for dataset 1
 \item d2vars - variable names for dataset 2
 \item fun - function name to process on these variables
 \item extraparams - extra parameters that need to be applied with the function.
 \item weights - a weight vector to scale each matrix (not used in processFunctionList).
}}

\item{giveWeight}{a logical parameter indicating whether or not the output
should be a list of weights or indices (default).}
}
\value{
something
}
\description{
Find duplicates in one or two data sets
}
\details{
this function will take in one or two data sets, a list of functions
  to apply to specific columns of the data set, and a threshold to determine
  what is a match. It will return a list from \code{\link{returnMatches}}
  where each element represents a different potential match. Within each
  element, there is a two-element list where each contains either indices or
  weights for each sample that matched below the threshold.
}
\examples{
indata <- system.file("data", package = "epimatch")
indata <- dir(indata, full.names = TRUE)
x <- lapply(indata, read.csv, stringsAsFactors = FALSE)

matchEpiData(dat1 = x[[1]],
                 dat2 = NULL,
                 funlist = list(
                   ID = list(d1vars = "Outbreak.ID.",
                             d2vars = NULL,
                             fun = "nameDists",
                             extraparams = NULL,
                             weight = 0.5),
                   names = list(d1vars = "Name..as.given.",
                                d2vars = NULL,
                                fun = "nameDists",
                                extraparams = NULL,
                                weight = 0.5)
                   ),
                 thresh = 0.5)
}

