% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/returnMatches.R
\name{returnMatches}
\alias{returnMatches}
\title{return matching indices}
\usage{
returnMatches(nRowD1, nRowD2, distMatrix, thresh)
}
\arguments{
\item{the}{number of rows for both datasets (ordered as they have been for all functions) and a distance matrix
thresh: distance threshold (smaller means more conservative matches)}
}
\value{
list where each index are the row indices for data rows in dataset 1 (d1),
and 2 (d2) that correspond to the same person
}
\description{
return matching indices
}

